fix: moving sweeper.go for fabric_cloud_router to internal/resources/cloud_router
